+++ modules/web/src/com/haulmont/cuba/web/gui/components/WebFileUploadField.java
@@ -165,8 +165,8 @@ public class WebFileUploadField extends WebAbstractComponent<UploadComponent> im
         impl.setDescription(null);
 
         Configuration configuration = AppBeans.get(Configuration.NAME);
-        final int maxUploadSizeMb = configuration.getConfig(ClientConfig.class).getMaxUploadSizeMb();
-        final int maxSizeBytes = maxUploadSizeMb * BYTES_IN_MEGABYTE;
+        final long maxUploadSizeMb = configuration.getConfig(ClientConfig.class).getMaxUploadSizeMb();
+        final long maxSizeBytes = maxUploadSizeMb * BYTES_IN_MEGABYTE;
 
         impl.setFileSizeLimit(maxSizeBytes);
