+++ modules/web/src/com/haulmont/cuba/web/toolkit/ui/client/fileupload/CubaFileUploadState.java
@@ -34,7 +34,7 @@ public class CubaFileUploadState extends TabIndexState {
     public String unableToUploadFileMessage;
 
     @NoLayout
-    public int fileSizeLimit;
+    public long fileSizeLimit;
 
     @NoLayout
     public Connector dropZone;
